feat: refactor and simplify AppStateProvider boilerplate for state stores
