+++ app/src/client/data-explorers/layouts/BaseDataExplorer/index.tsx
@@ -1,26 +1,29 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { RouterProvider } from 'react-router-dom';
 
-import type { DataExplorersStore, GenericStateStore } from '@nlpssa-app-types/common/main';
+import type {
+    DataExplorersStore,
+    GenericStateStore,
+} from '@nlpssa-app-types/common/main';
 import { BasePage } from 'client/common/layouts';
-import AppStateProvider from 'client/data-explorers/store/AppStateProvider';
+import { BaseStateContext, BaseDispatchContext } from 'client/common/store';
 import browserRouter from 'client/data-explorers/routes';
 
-function BaseDataExplorer({
-    initialPageData, // force formatting
-}: {
-    initialPageData: GenericStateStore<DataExplorersStore>;
-}) {
-    console.log('data-explorers - BaseDataExplorer', { initialPageData });
+function BaseDataExplorer() {
+    const state = useContext(BaseStateContext);
+    const dispatch = useContext(BaseDispatchContext);
+
+    const { user } = state as GenericStateStore<DataExplorersStore>;
+
+    console.log(`[data-explorers - ${BaseDataExplorer.name}]\n`, { state });
 
     return (
         <BasePage // force formatting
+            appDispatch={dispatch}
             pageTitle={`Data Explorers: <something-dynamic-here>`}
-            userData={initialPageData?.user}
+            userData={user}
         >
-            <AppStateProvider initialState={initialPageData}>
-                <RouterProvider router={browserRouter} />
-            </AppStateProvider>
+            <RouterProvider router={browserRouter} />
         </BasePage>
     );
 }
