+++ app/src/client/data-explorers/store/index.ts
@@ -1,6 +1,8 @@
-import AppStateProvider from './AppStateProvider';
+import { AppStateProviderHOC } from 'client/common/store';
 import * as articleDataBySlugActions from './actions';
 import { default as reducer } from './reducer';
+
+const AppStateProvider = AppStateProviderHOC(reducer);
 
 const actions = {
     ...articleDataBySlugActions,
